# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: HomeStock
  description: '

    A self-hosted home inventory system.


    Every response uses the same envelope: `{"data": ..., "error": null}`.

    On failure `data` is null and `error` holds a code and a message.


    The server runs every AI model on the CPU. There is no GPU. The AI routes

    return a job instead of a result when inference takes longer than

    `HS_AI_INLINE_TIMEOUT` seconds.

    '
  version: 0.1.0
paths:
...
  /api/sync/changes:
    get:
      tags:
      - Sync
      summary: Return everything that changed since a timestamp.
      description: Omit ``since`` on the first sync to receive the whole dataset. Send back the ``server_time``
        value from the reply on the next call. Deleted rows appear in the ``deleted`` map, because a hard
        delete could never reach the client.
      operationId: get_changes_api_sync_changes_get
      security:
      - HTTPBearer: []
      parameters:
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: An ISO 8601 timestamp from an earlier reply.
          title: Since
        description: An ISO 8601 timestamp from an earlier reply.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 5000
          minimum: 1
          default: 1000
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_SyncChanges_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sync/push:
    post:
      tags:
      - Sync
      summary: Upload the changes that the client made while it was offline.
      description: The server wins every conflict. A rejected change appears in ``conflicts`` so that
        the client can tell the user.
      operationId: push_changes_api_sync_push_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPush'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_SyncPushResult_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
components:
  schemas:
...
    CableRow:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
        connector_a:
          anyOf:
          - type: string
          - type: 'null'
          title: Connector A
        connector_b:
          anyOf:
          - type: string
          - type: 'null'
          title: Connector B
        length_cm:
          anyOf:
          - type: integer
          - type: 'null'
          title: Length Cm
        colour:
          anyOf:
          - type: string
          - type: 'null'
          title: Colour
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        specification:
          anyOf:
          - type: string
          - type: 'null'
          title: Specification
        quantity:
          type: integer
          title: Quantity
        price:
          anyOf:
          - type: string
          - type: 'null'
          title: Price
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Item Id
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - name
      - kind
      - connector_a
      - connector_b
      - length_cm
      - colour
      - brand
      - specification
      - quantity
      - price
      - notes
      - location_id
      - item_id
      - version
      - created_at
      - updated_at
      title: CableRow
      description: One cable, exactly as the table holds it. The phone reads this.
...
    ComponentRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        model_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Number
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        default_price:
          anyOf:
          - type: string
          - type: 'null'
          title: Default Price
        is_consumable:
          type: boolean
          title: Is Consumable
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        thumbnail_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Path
          description: The picture that stands for this component.
        photo_count:
          type: integer
          title: Photo Count
          default: 0
      type: object
      required:
      - id
      - user_id
      - name
      - brand
      - model_number
      - category
      - description
      - default_price
      - is_consumable
      - version
      - created_at
      - updated_at
      title: ComponentRead
      description: One catalogue entry.
...
    CustomFieldRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        item_id:
          type: string
          format: uuid
          title: Item Id
        field_name:
          type: string
          title: Field Name
        field_value:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Value
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - item_id
      - field_name
      - field_value
      - version
      - created_at
      - updated_at
      title: CustomFieldRead
      description: A custom field.
...
    Envelope_SyncChanges_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/SyncChanges'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[SyncChanges]
    Envelope_SyncPushResult_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/SyncPushResult'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[SyncPushResult]
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          description: A stable machine readable code.
        message:
          type: string
          title: Message
          description: A message for a person to read.
        details:
          anyOf:
          - type: object
          - type: 'null'
          title: Details
          description: Extra context, such as field errors.
      type: object
      required:
      - code
      - message
      title: ErrorDetail
      description: The error half of the response envelope.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ItemComponentRow:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        item_id:
          type: string
          format: uuid
          title: Item Id
        component_id:
          type: string
          format: uuid
          title: Component Id
        quantity:
          type: integer
          title: Quantity
        price:
          anyOf:
          - type: string
          - type: 'null'
          title: Price
        serial_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Serial Number
        fitted_on:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Fitted On
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - item_id
      - component_id
      - quantity
      - price
      - serial_number
      - fitted_on
      - notes
      - version
      - created_at
      - updated_at
      title: ItemComponentRow
      description: 'One fitted component, exactly as the table holds it.


        The sync endpoint sends this. The phone holds the catalogue too, so it

        joins the name itself.'
...
    ItemPhotoRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        item_id:
          type: string
          format: uuid
          title: Item Id
        file_path:
          type: string
          title: File Path
        thumbnail_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Path
        is_primary:
          type: boolean
          title: Is Primary
        ai_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Description
        ocr_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Ocr Text
          description: The text that OCR read from the photograph, if any.
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - item_id
      - file_path
      - thumbnail_path
      - is_primary
      - ai_description
      - created_at
      title: ItemPhotoRead
      description: One photograph of an item.
    ItemRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        subcategory:
          anyOf:
          - type: string
          - type: 'null'
          title: Subcategory
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        serial_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Serial Number
        barcode:
          anyOf:
          - type: string
          - type: 'null'
          title: Barcode
        purchase_price:
          anyOf:
          - type: string
          - type: 'null'
          title: Purchase Price
        current_value:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Value
        purchase_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Purchase Date
        purchase_location:
          anyOf:
          - type: string
          - type: 'null'
          title: Purchase Location
        warranty_expires:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Warranty Expires
        condition:
          anyOf:
          - type: string
          - type: 'null'
          title: Condition
        quantity:
          type: integer
          title: Quantity
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
        is_lent:
          type: boolean
          title: Is Lent
        lent_to:
          anyOf:
          - type: string
          - type: 'null'
          title: Lent To
        lent_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Lent Date
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        tags:
          items:
            $ref: '#/components/schemas/TagRead'
          type: array
          title: Tags
        primary_photo:
          anyOf:
          - $ref: '#/components/schemas/ItemPhotoRead'
          - type: 'null'
      type: object
      required:
      - id
      - user_id
      - location_id
      - name
      - description
      - category
      - subcategory
      - brand
      - model
      - serial_number
      - barcode
      - purchase_price
      - current_value
      - purchase_date
      - purchase_location
      - warranty_expires
      - condition
      - quantity
      - notes
      - is_lent
      - lent_to
      - lent_date
      - version
      - created_at
      - updated_at
      title: ItemRead
      description: An item in a list. It carries no nested collections except tags.
...
    LocationRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        parent_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Id
        name:
          type: string
          title: Name
        type:
          type: string
          title: Type
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        photo_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Path
        sort_order:
          type: integer
          title: Sort Order
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - parent_id
      - name
      - type
      - description
      - photo_path
      - sort_order
      - version
      - created_at
      - updated_at
      title: LocationRead
      description: A location, without its children.
...
    MaintenanceRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        item_id:
          type: string
          format: uuid
          title: Item Id
        description:
          type: string
          title: Description
        date_performed:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date Performed
        next_due_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Next Due Date
        cost:
          anyOf:
          - type: string
          - type: 'null'
          title: Cost
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - item_id
      - description
      - date_performed
      - next_due_date
      - cost
      - notes
      - version
      - created_at
      - updated_at
      title: MaintenanceRead
      description: A maintenance record.
...
    ReceiptRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        file_path:
          type: string
          title: File Path
        thumbnail_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Path
        vendor:
          anyOf:
          - type: string
          - type: 'null'
          title: Vendor
        purchase_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Purchase Date
        total_amount:
          anyOf:
          - type: string
          - type: 'null'
          title: Total Amount
        currency:
          type: string
          title: Currency
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - file_path
      - thumbnail_path
      - vendor
      - purchase_date
      - total_amount
      - currency
      - version
      - created_at
      - updated_at
      title: ReceiptRead
      description: A receipt in a list.
...
    SpareRow:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        component_id:
          type: string
          format: uuid
          title: Component Id
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        quantity:
          type: integer
          title: Quantity
        minimum_quantity:
          type: integer
          title: Minimum Quantity
        unit_price:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit Price
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - component_id
      - location_id
      - quantity
      - minimum_quantity
      - unit_price
      - notes
      - version
      - created_at
      - updated_at
      title: SpareRow
      description: One row of stock, exactly as the table holds it.
...
    SyncChange:
      properties:
        entity:
          $ref: '#/components/schemas/SyncEntity'
        op:
          $ref: '#/components/schemas/SyncOp'
        id:
          type: string
          format: uuid
          title: Id
          description: The client creates the UUID, so a create is repeatable.
        base_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Base Version
          description: The version the client last saw. Used to detect a conflict.
        payload:
          anyOf:
          - type: object
          - type: 'null'
          title: Payload
          description: The row fields. Null for a delete.
        client_updated_at:
          type: string
          format: date-time
          title: Client Updated At
      type: object
      required:
      - entity
      - op
      - id
      - client_updated_at
      title: SyncChange
      description: One change that the client made while it was offline.
    SyncChanges:
      properties:
        since:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Since
        server_time:
          type: string
          format: date-time
          title: Server Time
          description: Send this value back as ``since`` on the next sync.
        items:
          items:
            $ref: '#/components/schemas/ItemRead'
          type: array
          title: Items
        item_photos:
          items:
            $ref: '#/components/schemas/ItemPhotoRead'
          type: array
          title: Item Photos
        locations:
          items:
            $ref: '#/components/schemas/LocationRead'
          type: array
          title: Locations
        tags:
          items:
            $ref: '#/components/schemas/TagRead'
          type: array
          title: Tags
        receipts:
          items:
            $ref: '#/components/schemas/ReceiptRead'
          type: array
          title: Receipts
        maintenance_logs:
          items:
            $ref: '#/components/schemas/MaintenanceRead'
          type: array
          title: Maintenance Logs
        custom_fields:
          items:
            $ref: '#/components/schemas/CustomFieldRead'
          type: array
          title: Custom Fields
        components:
          items:
            $ref: '#/components/schemas/ComponentRead'
          type: array
          title: Components
        item_components:
          items:
            $ref: '#/components/schemas/ItemComponentRow'
          type: array
          title: Item Components
        spares:
          items:
            $ref: '#/components/schemas/SpareRow'
          type: array
          title: Spares
        cables:
          items:
            $ref: '#/components/schemas/CableRow'
          type: array
          title: Cables
        photos:
          items:
            $ref: '#/components/schemas/PhotoRow'
          type: array
          title: Photos
        activity:
          items:
            $ref: '#/components/schemas/ActivityRow'
          type: array
          title: Activity
        deleted:
          additionalProperties:
            items:
              type: string
              format: uuid
            type: array
          propertyNames:
            $ref: '#/components/schemas/SyncEntity'
          type: object
          title: Deleted
        has_more:
          type: boolean
          title: Has More
          description: True when the page limit cut the result. Sync again.
          default: false
      type: object
      required:
      - since
      - server_time
      title: SyncChanges
      description: 'Everything that changed on the server since a timestamp.


        ``deleted`` holds the ids of soft deleted rows. A hard delete cannot

        appear in a delta query, so every user table keeps a ``deleted_at``

        tombstone instead.'
...
    SyncEntity:
      type: string
      enum:
      - item
      - item_photo
      - location
      - tag
      - receipt
      - maintenance
      - custom_field
      - component
      - item_component
      - spare
      - cable
      - photo
      - activity
      title: SyncEntity
      description: The tables that the mobile application mirrors.
...
    SyncPush:
      properties:
        changes:
          items:
            $ref: '#/components/schemas/SyncChange'
          type: array
          maxItems: 500
          minItems: 1
          title: Changes
      type: object
      required:
      - changes
      title: SyncPush
      description: A batch of offline changes.
    SyncPushResult:
      properties:
        applied:
          type: integer
          title: Applied
        rejected:
          type: integer
          title: Rejected
        conflicts:
          items:
            $ref: '#/components/schemas/SyncConflict'
          type: array
          title: Conflicts
        errors:
          items:
            $ref: '#/components/schemas/SyncError'
          type: array
          title: Errors
          description: Changes that failed. The client keeps these and retries.
        server_time:
          type: string
          format: date-time
          title: Server Time
      type: object
      required:
      - applied
      - rejected
      - server_time
      title: SyncPushResult
      description: What the server did with a pushed batch.
...
    TagRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        color:
          type: string
          title: Color
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - name
      - color
      - created_at
      - updated_at
      title: TagRead
      description: A tag.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      description: A JWT access token.
      scheme: bearer